break up povw commands in modules
