+++ crates/boundless-cli/src/commands/povw/send_update.rs
@@ -1,0 +1,187 @@
+// Copyright 2025 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::path::PathBuf;
+
+use alloy::{
+    primitives::Address,
+    providers::{Provider, ProviderBuilder},
+    signers::local::PrivateKeySigner,
+};
+use anyhow::{bail, ensure, Context};
+use boundless_povw_guests::log_updater::{prover::LogUpdaterProver, IPovwAccounting};
+use clap::Args;
+use risc0_povw::guest::Journal as LogBuilderJournal;
+use risc0_zkvm::{default_prover, ProverOpts};
+
+use super::State;
+use crate::config::GlobalConfig;
+
+/// Send a work log update to the PoVW accounting contract.
+///
+/// To prepare the update, this command creates a Groth16 proof, compressing the updates to be sent
+/// and proving that they are authorized by the signing key for the work log.
+#[non_exhaustive]
+#[derive(Args, Clone, Debug)]
+pub struct PovwSendUpdate {
+    /// State of the work log, including receipts produced by the prove-update command.
+    #[arg(short, long)]
+    pub state: PathBuf,
+
+    /// Private key used to sign work log updates. Should have an address equal to the work log ID.
+    ///
+    /// If this option is not set, the value of the private key from global config will be used.
+    #[clap(long, env = "WORK_LOG_PRIVATE_KEY", hide_env_values = true)]
+    pub work_log_private_key: Option<PrivateKeySigner>,
+
+    // TODO(povw): Provide a default here, similar to the Deployment struct in boundless-market.
+    /// Address of the PoVW accounting contract.
+    pub povw_accounting_address: Address,
+}
+
+impl PovwSendUpdate {
+    /// Run the [PovwSendUpdate] command.
+    pub async fn run(&self, global_config: &GlobalConfig) -> anyhow::Result<()> {
+        let tx_signer = global_config.require_private_key()?;
+        let work_log_signer = self.work_log_private_key.as_ref().unwrap_or(&tx_signer);
+        let rpc_url = global_config.require_rpc_url()?;
+
+        // Load the state and check to make sure the private key matches.
+        let state = State::load(&self.state)
+            .with_context(|| format!("Failed to load state from {}", self.state.display()))?;
+        ensure!(
+            Address::from(state.log_id) == work_log_signer.address(),
+            "Signer does not match the state log ID: signer: {}, state: {}",
+            work_log_signer.address(),
+            state.log_id
+        );
+
+        // Connect to the chain.
+        let provider = ProviderBuilder::new()
+            .wallet(tx_signer.clone())
+            .connect(rpc_url.as_str())
+            .await
+            .with_context(|| format!("failed to connect provider to {rpc_url}"))?;
+
+        let chain_id = provider
+            .get_chain_id()
+            .await
+            .with_context(|| format!("failed to get chain ID from {rpc_url}"))?;
+        let povw_accounting = IPovwAccounting::new(self.povw_accounting_address, provider.clone());
+
+        // Get the current work log commit, to determine which update(s) should be applied.
+        let onchain_commit =
+            povw_accounting.workLogCommit(state.log_id.into()).call().await.with_context(|| {
+                format!(
+                    "Failed to get work log commit for {:x} from {:x}",
+                    state.log_id, self.povw_accounting_address
+                )
+            })?;
+
+        // Check if the latest log builder receipt has an updated_commit value equal to what is
+        // onchain. If so, the onchain work log is already up to date.
+        let Some(latest_receipt) = state.log_builder_receipts.last() else {
+            bail!("Loaded state has no log builder receipts")
+        };
+        let latest_receipt_journal = LogBuilderJournal::decode(&latest_receipt.journal.bytes)
+            .context("Failed to decode journal from latest receipt")?;
+        if bytemuck::cast::<_, [u8; 32]>(latest_receipt_journal.updated_commit) == *onchain_commit {
+            tracing::info!("Onchain PoVW accounting contract is already up to date with the latest commit in state");
+            return Ok(());
+        }
+
+        // Find the index of the receipt in the state that has an initial commit equal to the
+        // commit current onchain. We will send all updates after that point.
+        let matching_receipt_index = state
+            .log_builder_receipts
+            .iter()
+            .enumerate()
+            .rev()
+            .map(|(i, receipt)| {
+                let journal =
+                    LogBuilderJournal::decode(&receipt.journal.bytes).with_context(|| {
+                        format!("Failed to decode journal from receipt in state at index {i}")
+                    })?;
+                anyhow::Ok(
+                    (bytemuck::cast::<_, [u8; 32]>(journal.initial_commit) == *onchain_commit)
+                        .then_some(i),
+                )
+            })
+            .find_map(|x| x.transpose())
+            .with_context(|| {
+                format!("Failed to find receipt with initial commit matching {onchain_commit}")
+            })??;
+
+        // Iterate over all the log builder receipts that should be sent to the chain.
+        // NOTE: In most cases, this will be one receipt. It may be more if the prover previously
+        // built a work log update but it failed to send (e.g. network instability or high gas
+        // fees caused the transaction not to go through).
+        let receipts_for_update = state.log_builder_receipts[matching_receipt_index..].to_vec();
+        if receipts_for_update.len() > 1 {
+            tracing::info!(
+                "Updating onchain work log {:x} with {} update receipts",
+                state.log_id,
+                receipts_for_update.len()
+            )
+        }
+        for receipt in receipts_for_update {
+            let prover = LogUpdaterProver::builder()
+                .prover(default_prover())
+                .chain_id(chain_id)
+                .contract_address(self.povw_accounting_address)
+                .prover_opts(ProverOpts::groth16())
+                .build()
+                .context("Failed to build prover for Log Updater")?;
+
+            // Sign and prove the authorized work log update.
+            // TODO(povw): Provide more info here.
+            tracing::info!("Proving work log update");
+            let prove_info = prover
+                .prove_update(receipt, work_log_signer)
+                .await
+                .context("Failed to prove authorized log update")?;
+
+            tracing::info!("Sending work log update transaction");
+            let tx_result = povw_accounting
+                .update_work_log(&prove_info.receipt)
+                .context("Failed to construct update transaction")?
+                .send()
+                .await
+                .context("Failed to send update transaction")?;
+            tracing::info!(tx_hash = %tx_result.tx_hash(), "Sent transaction for work log update");
+
+            let tx_receipt = tx_result
+                .get_receipt()
+                .await
+                .context("Failed to receive receipt for update transaction")?;
+
+            // Extract the WorkLogUpdated event
+            let work_log_updated_event = tx_receipt
+                .logs()
+                .iter()
+                .filter_map(|log| log.log_decode::<IPovwAccounting::WorkLogUpdated>().ok())
+                .next();
+
+            if let Some(event) = work_log_updated_event {
+                let data = event.inner.data;
+                tracing::info!(updated_commit = %data.updatedCommit, update_value = data.updateValue.to::<u64>(), "Work log update confirmed");
+            } else {
+                tracing::info!("Work log update confirmed");
+                tracing::warn!("No WorkLogUpdated event in transaction receipt");
+            }
+        }
+
+        Ok(())
+    }
+}
